Ensure mapAsync completion is only handled by polling
Trying to resolve random unexplained crashes of the sort:

-[MTLDebugDevice notifyExternalReferencesNonZeroOnDealloc:]:3357: failed assertion `The following Metal object is being destroyed while still required to be alive by the command buffer 0x148ce7900 (label: (wgpu internal) Signal):
<MTLToolsObject: 0x108bb2330> -> <AGXG14XFamilyBuffer: 0x10f2acb00>
    label =
    length = 399360
    cpuCacheMode = MTLCPUCacheModeDefaultCache
    storageMode = MTLStorageModeShared

This fix is based on a guess that's probably not right, but the fix is otherwise harmless and better specifies the behavior we want anyway.

The guess is that maybe this crash is occurring if the completion callback is unintentionally handled on a secondary thread instead of by the polling (and that maybe then the polling doesn't wait forever because of some other reason, maybe a previous completed event).

I couldn't find documentation for the behavior when mode is unspecified (inited to zero which isn't a listed valid mode). We're only interested in the synchronous completion mode handled by the polling, so define that as the mode.